Add detailed write-up and images for university project
Expanded the university project page with a  write-up of the Data Science Individual Project analysing playstyle in Counter-Strike, including methodology, analysis, and interpretations. Added two illustrative SVG images to support the content.

TODO: Add all the other modules, maybe move the final project to its own project and link to it instead of making it collapsable :/